booking-service/requirements.txt
apifairy==1.3.0
apispec==6.3.0
bidict==0.22.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
paho-mqtt==1.5.1
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.0.0
webargs==8.3.0
Werkzeug==3.0.1
wsproto==1.2.0
pymongo==3.12.0
flask_jwt_extended==4.5.3
bcrypt==3.2.1
python-dotenv==1.0.0
dnspython==2.4.2
timedelta
